Transition animation built into the NavigationModal component. After checking on android/ios and tablet whether everything functions and displays correctly, this branch will be merged with main.
